Overflows-Apis/requirements.txt
# ========= CORE FRAMEWORK DEPENDENCIES =========
fastapi==0.117.1
uvicorn==0.36.0
pydantic==2.11.9
pydantic_core==2.33.2
starlette==0.48.0

# ========= HTTP & NETWORKING =========
requests==2.32.5
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0
urllib3==2.5.0
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10

# ========= ASYNC & CONCURRENCY =========
anyio==4.10.0
sniffio==1.3.1

# ========= DATA PROCESSING & SCIENTIFIC COMPUTING =========
numpy==2.0.2
pillow==11.0.0

# ========= DATABASE & STORAGE =========
pymongo==4.15.3
dnspython==2.7.0

# ========= AWS & CLOUD SERVICES =========
boto3==1.40.55
botocore==1.40.55
s3transfer==0.14.0
jmespath==1.0.1

# ========= ENVIRONMENT & CONFIG =========
python-dotenv==1.1.1
click==8.3.0

# ========= AI & ML SERVICES =========
google-generativeai
replicate>=1.0.0

# ========= FILE PROCESSING =========
python-multipart
python-dateutil==2.9.0.post0
six==1.17.0
PyPDF2>=3.0.0
python-docx>=1.0.0
textract>=1.6.5

# ========= TYPE HINTS & ANNOTATIONS =========
typing_extensions==4.15.0
typing-inspection==0.4.1
annotated-types==0.7.0

# ========= CUSTOM/INTERNAL =========
synexa==0.1.1
